feat: 카드 레이아웃 컴포넌트 화
