+++ src/components/common/Card/CardList/PbCardList.tsx
@@ -1,8 +1,8 @@
 "use client";
 import React from "react";
-import PbCardItem from "@/components/common/CardItem/PbCardItem";
+import PbCardItem from "@/components/common/Card/CardItem/PbCardItem";
 
-function PbCardList(props: any) {
+function PbCardList({props}: any) {
   
   return (
     <ul>
